Cleaning of TitleResponse so we just use id rather than netflixid and also separating out TitleType
